First working copy
This is now working after some trial and error. Several changes to the
code and structure.

1. The buttons render correctly. I adapted this from a sample that used
a lot of absolute positioning, and didn't take into account rendering
anything else on the page. The buttons now render correctly.

2. out/ directory is now gone. gamepad viewer doesn't link to any raw
files from a GitHub repository directly, but will from a Gist. So the
out directory no longer makes any sense.

3. The buttons now have a darker ring when actuated. It looked fine up
close, but it was difficult to understand what was being pressed when
the stick wasn't at full resolution.

4. Override some defaults from gamepad viewer that interfered with the
buttons.